--- feature/center/src/jsMain/kotlin/com/mifos/feature/center/centerList/ui/CenterListScreen.js.kt
@@ -1,24 +1,0 @@
-/*
- * Copyright 2026 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See https://github.com/openMF/android-client/blob/master/LICENSE.md
- */
-package com.mifos.feature.center.centerList.ui
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-actual fun CenterListContent(
-    state: CenterListUiState,
-    isInSelectionMode: Boolean,
-    selectedItems: SelectedItemsState,
-    onRefresh: () -> Unit,
-    onCenterSelect: (Int) -> Unit,
-    modifier: Modifier,
-) {
-}
